+++ src/services/traits/InstallServiceHelperTrait.php
@@ -7,33 +7,9 @@
 
 namespace twentyfourhoursmedia\poll\services\traits;
 use Craft;
-use craft\models\FieldGroup;
 
 trait InstallServiceHelperTrait
 {
-
-
-    /**
-     * Gets a field group to install fields in (defaults to first)
-     *
-     * @param string|null $name
-     * @return FieldGroup | null
-     * @throws \Throwable
-     */
-    protected function enforceFieldGroupWithName(string $name = null)
-    {
-
-        $groups = Craft::$app->fields->getAllGroups();
-        foreach ($groups as $group) {
-            if (strtolower($group->name) === strtolower($name)) {
-                return $group;
-            }
-        }
-        $group = new FieldGroup();
-        $group->name = $name;
-        $success = Craft::$app->fields->saveGroup($group, true);
-        return $success ? $group : false;
-    }
 
     /**
      * @param string $handle
@@ -51,7 +27,7 @@ trait InstallServiceHelperTrait
      */
     protected function hasSectionWithHandle(string $handle): bool
     {
-        $section = Craft::$app->sections->getSectionByHandle($handle);
+        $section = Craft::$app->entries->getSectionByHandle($handle);
         return $section ? true : false;
     }
 
@@ -59,8 +35,8 @@ trait InstallServiceHelperTrait
      * Makes sure a field with a handle exists, if not retrieves the field from the callback and create it
      *
      * @param $handle
-     * @param callable $createCallback                         callback that should return a new field if it does not exist
-     * @param callable $createdCallback                        callback after field is created, wi
+     * @param callable $createCallback     callback that should return a new field if it does not exist
+     * @param callable $createdCallback    callback after field is created
      * @return bool|\craft\base\FieldInterface|null
      * @throws \Throwable
      */